pyproject.toml
[project]
name = "team-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi-admin>=1.0.4",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "fastapi[standard]>=0.124.0",
    "psycopg[binary]>=3.3.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
